+++ src/main/java/com/ldh/todolist/dao/CategoryDao.java
@@ -17,12 +17,12 @@ public interface CategoryDao {
 	void update(CategoryDto categoryDto);
 
 	//카테고리 삭제
-	void delete(int categoryNo);
+	void delete(Long categoryNo);
 	
 	//카테고리 조회
-	CategoryDto findById(int categoryNo);
+	CategoryDto findById(Long categoryNo);
 	
 	//카테고리 리스트 조회
-	List<CategoryDto> findByUsersNo(int usersNo);
+	List<CategoryDto> findByUsersNo(Long usersNo);
 	
 }
