+++ src/main/java/com/ldh/todolist/service/CategoryService.java
@@ -13,12 +13,12 @@ public interface CategoryService {
 	void updateCategory(CategoryDto categoryDto);
 
 	//카테고리 삭제
-	void deleteCategory(int categoryNo);
+	void deleteCategory(Long categoryNo);
 	
 	//카테고리 조회
-	CategoryDto findById(int categoryNo);
+	CategoryDto findById(Long categoryNo);
 	
 	//카테고리 목록 조회
-	List<CategoryDto> findByUsersNo(int usersNo);
+	List<CategoryDto> findByUsersNo(Long usersNo);
 
 }
